feat: add test, build, and import query param args.
