+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
       - name: Setup Node
         uses: actions/setup-node@v3.6.0
         with:
-          node-version: '20.3.1'
+          node-version: '20.4.0'
       - name: Install Dependencies
         run: npm ci
       - name: Save error log
@@ -29,5 +29,11 @@ jobs:
           path: npm-debug.log
       - name: Lint
         run: npm run lint
+      - name: Test
+        run: npm test
+      - name: Report Coverage
+        uses: codecov/codecov-action@v3.1.4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
       - name: Pack
         run: npm pack
