+++ .github/workflows/publish.yml
@@ -14,7 +14,7 @@ jobs:
       - name: Setup Node
         uses: actions/setup-node@v3.6.0
         with:
-          node-version: '20.3.1'
+          node-version: '20.4.0'
       - name: Install Dependencies
         run: npm ci
       - name: Save error log
@@ -25,6 +25,8 @@ jobs:
           path: npm-debug.log
       - name: Lint
         run: npm run lint
+      - name: Test
+        run: npm test
       - name: Pack
         run: npm pack
       - name: Push to NPM registry
