+++ codecov.yml
@@ -1,0 +1,10 @@
+coverage:
+  status:
+    project:
+      default:
+        target: 80%
+        threshold: 0
+    patch:
+      default:
+        target: 81%
+        threshold: 1%
